% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinome_data}
\alias{kinome_data}
\title{A list object that contains kinome data for several model organisms}
\format{
## `kinome_data`
A list where the first level represents one species each.
}
\source{
Contains data from several publications by Manning et al. who kindly allowed the redistribution of the data.
}
\usage{
kinome_data
}
\description{
A list object that contains kinome data for several model organisms
}
\keyword{datasets}
