% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prokvis2-package.R
\docType{package}
\name{prokvis2-package}
\alias{prokvis2}
\alias{prokvis2-package}
\title{prokvis2: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yasche/prokvis2}
  \item Report bugs at \url{https://github.com/yasche/prokvis2/issues}
}

}
\author{
\strong{Maintainer}: Yannik Schermer \email{your@email.com} [copyright holder]

}
\keyword{internal}
